style(new code style): cut the viewer.cpp, divide to other files shorten the angle-relative file [   changes ] C: viewer.cpp, A: pointcloud.{h, inc} [ reference ] [  computer ]
